requirements.txt
streamlit==1.28.0
langchain==0.1.0
pypdf==4.0.0
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
python-dotenv==1.0.0
groq
huggingface_hub>=0.20.0
tokenizers>=0.15.2
transformers>=4.37.0
httpx==0.27.0